+++ jetty-client/src/main/scala/org/http4s/jetty/client/ResponseListener.scala
@@ -24,11 +24,10 @@ import cats.effect.std.Queue
 import cats.syntax.all._
 import fs2.Stream._
 import fs2._
-import org.eclipse.jetty.client.api.Result
-import org.eclipse.jetty.client.api.{Response => JettyResponse}
+import org.eclipse.jetty.client.Result
+import org.eclipse.jetty.client.{Response => JettyResponse}
 import org.eclipse.jetty.http.HttpFields
 import org.eclipse.jetty.http.{HttpVersion => JHttpVersion}
-import org.eclipse.jetty.util.{Callback => JettyCallback}
 import org.http4s.internal.CollectionCompat.CollectionConverters._
 import org.http4s.jetty.client.ResponseListener.Item
 import org.http4s.jetty.client.internal.loggingAsyncCallback
@@ -41,7 +40,7 @@ private[jetty] final case class ResponseListener[F[_]](
     cb: Callback[Resource[F, Response[F]]],
     dispatcher: Dispatcher[F],
 )(implicit F: Async[F])
-    extends JettyResponse.Listener.Adapter {
+    extends JettyResponse.Listener {
   import ResponseListener.logger
 
   /* Needed to properly propagate client errors */
@@ -89,14 +88,14 @@ private[jetty] final case class ResponseListener[F[_]](
   override def onContent(
       response: JettyResponse,
       content: ByteBuffer,
-      callback: JettyCallback,
   ): Unit = {
     val copy = ByteBuffer.allocate(content.remaining())
     copy.put(content).flip()
-    enqueue(Item.Buf(copy)) {
-      case Right(_) => F.delay(callback.succeeded())
+    enqueueSync(Item.Buf(copy)) {
+      case Right(_) =>
+        F.unit
       case Left(e) =>
-        F.delay(logger.error(e)("Error in asynchronous callback")) >> F.delay(callback.failed(e))
+        F.delay(logger.error(e)("Error in asynchronous callback"))
     }
   }
 
@@ -115,17 +114,31 @@ private[jetty] final case class ResponseListener[F[_]](
   // (the request might complete after the response has been entirely received)
   override def onComplete(result: Result): Unit = ()
 
-  private def abort(t: Throwable, response: JettyResponse): Unit =
-    if (!response.abort(t)) // this also aborts the request
-      logger.error(t)("Failed to abort the response")
-    else
-      closeStream()
+  private def abort(t: Throwable, response: JettyResponse): Unit = {
+    import scala.compat.java8.FutureConverters._
+
+    dispatcher.unsafeRunAndForget(
+      Async[F]
+        .fromFuture(F.delay(response.abort(t).toScala))
+        .map { aborted =>
+          if (!aborted)
+            logger.error(t)("Failed to abort the response")
+          else
+            closeStream()
+        }
+        .attempt
+        .flatMap(loggingAsyncCallback[F, Unit](logger))
+    )
+  }
 
   private def closeStream(): Unit =
     enqueue(Item.Done)(loggingAsyncCallback[F, Unit](logger))
 
   private def enqueue(item: Item)(cb: Either[Throwable, Unit] => F[Unit]): Unit =
     dispatcher.unsafeRunAndForget(queue.offer(item.some).attempt.flatMap(cb))
+
+  private def enqueueSync(item: Item)(cb: Either[Throwable, Unit] => F[Unit]): Unit =
+    dispatcher.unsafeRunSync(queue.offer(item.some).attempt.flatMap(cb))
 }
 
 private[jetty] object ResponseListener {
